refactoring project importers
