+++ src/main/java/org/earthtime/projects/projectImporters/AbstractProjectImporterFromLegacyCSVFile.java
@@ -182,9 +182,4 @@ public abstract class AbstractProjectImporterFromLegacyCSVFile {
         this.mruFolder = mruFolder;
     }
 
-    /**
-     *
-     */
-    public static void writeAndOpenCSVFileOfLegacyDataSampleFieldNames () {
-    }
 }
